Predict Refactor
