+++ morgana/cli/classifying_unet.py
@@ -10,7 +10,7 @@ image_folder = (
     "/Users/nicholb/Documents/data/organoid_data/seperatedStacks/image_1_MMStack_control_DMSO_1-1.ome_restacked/ROI2"
 )
 flist_in = ioDT.get_image_list(image_folder, string_filter="_GT", mode_filter="exclude")
-classifier, scaler, params = ioML.load_model(model_folder, model="MLP")
+classifier, scaler, params = ioML.load_model(model_folder)
 for f in flist_in:
     img = imread(f)
     pred, prob = predict.predict_image_unet(img, classifier, scaler, image_size=(512, 512))
